cleaner su in launcher script
The `-` option is there primarily for one side effect: it does`cd` to
riak user's HOME (/var/lib/riak).  Critically, this dir has the
erlang cookie file, which is read from the new node started as part of
some riak-admin commands (e.g., riak-admin test). It, however, creates
a barrier for env vars, which is undesirable.

If we cd manually into riak's HOME before the call to `su`, the piggy
node now can find the cookie in its cwd, and execution of the command
succeeds.